modification et suppression en fonction des autorisation

+++ controllers/comment.js
@@ -1,4 +1,5 @@
-const db = require('../models')
+const db = require('../models');
+const verify = require('../middleware/verify');
 const Comment = db.comment;
 
 exports.create = async (req, res, next) => {
@@ -7,7 +8,7 @@ exports.create = async (req, res, next) => {
         await Comment.create({ ...req.body });
         return res.status(201).json({ message: 'Commentaire créé !' });
     }
-    catch(err) {
+    catch (err) {
         res.status(400).json({ err });
         console.error(err);
     }
@@ -26,7 +27,7 @@ exports.getAll = async (req, res, next) => {
 
 exports.get = async (req, res, next) => {
     try {
-        const result = await Comment.findOne({ include: ['user'], where: { post_id: req.params.id, id: req.params.idcom  } });
+        const result = await Comment.findOne({ include: ['user'], where: { post_id: req.params.id, id: req.params.idcom } });
         return res.status(200).json(result);
     }
     catch (err) {
@@ -37,8 +38,15 @@ exports.get = async (req, res, next) => {
 
 exports.update = async (req, res, next) => {
     try {
-        await Comment.update({ content: req.body.content }, { where: { post_id: req.params.id, id: req.params.idcom } });
-        return res.status(200).json({ message: 'Commentaire mis à jour !' })
+        const userId = verify.verifyUser(req, res, next);
+        const result = await Comment.findOne({ where: { post_id: req.params.id, id: req.params.idcom }, raw: true });
+        if (userId == result.user_id) {
+            await Comment.update({ content: req.body.content }, { where: { post_id: req.params.id, id: req.params.idcom } });
+            return res.status(200).json({ message: 'Commentaire mis à jour !' });
+        }
+        else {
+            res.status(401).json({ message: 'Utilisateur non autorisé !' });
+        }
     }
     catch (err) {
         res.status(400).json({ err });
@@ -48,8 +56,15 @@ exports.update = async (req, res, next) => {
 
 exports.delete = async (req, res, next) => {
     try {
-        await Comment.destroy({ where: { post_id: req.params.id, id: req.params.idcom } });
-        return res.status(200).json({ message: 'Commentaire supprimé !' })
+        const userId = verify.verifyUser(req, res, next);
+        const result = await Comment.findOne({ where: { post_id: req.params.id, id: req.params.idcom }, raw: true });
+        if (userId == result.user_id) {
+            await Comment.destroy({ where: { post_id: req.params.id, id: req.params.idcom } });
+            return res.status(200).json({ message: 'Commentaire supprimé !' });
+        }
+        else {
+            res.status(401).json({ message: 'Utilisateur non autorisé !' });
+        }
     }
     catch (err) {
         res.status(400).json({ err });
