+++ controllers/post.js
@@ -1,6 +1,8 @@
-const db = require('../models')
+const db = require('../models');
+const verify = require('../middleware/verify');
 const Post = db.post;
 const Comment = db.comment;
+const Like = db.like;
 
 exports.create = async (req, res, next) => {
     try {
@@ -9,7 +11,7 @@ exports.create = async (req, res, next) => {
         });
         return res.status(201).json({ message: 'Publication créée !' });
     }
-    catch(err) {
+    catch (err) {
         res.status(400).json({ err });
         console.error(err);
     }
@@ -28,7 +30,7 @@ exports.getAll = async (req, res, next) => {
 
 exports.get = async (req, res, next) => {
     try {
-        const result = await Post.findOne({ include: ['user'] , where: { id: req.params.id } });
+        const result = await Post.findOne({ include: ['user'], where: { id: req.params.id } });
         return res.status(200).json(result);
     }
     catch (err) {
@@ -39,12 +41,15 @@ exports.get = async (req, res, next) => {
 
 exports.update = async (req, res, next) => {
     try {
-        await Post.update({
-            title: req.body.title,
-            content: req.body.content
-        },
-            { where: { id: req.params.id } });
-        return res.status(200).json({ message: 'Publication mise à jour !' })
+        const userId = verify.verifyUser(req, res, next);
+        const result = await Post.findOne({ where: { id: req.params.id }, raw: true });
+        if (userId == result.user_id) {
+            await Post.update({ ...req.body }, { where: { id: req.params.id } });
+            return res.status(200).json({ message: 'Publication mise à jour !' })
+        }
+        else {
+            res.status(401).json({ message: 'Utilisateur non autorisé !' });
+        }
     }
     catch (err) {
         res.status(400).json({ err });
@@ -54,9 +59,17 @@ exports.update = async (req, res, next) => {
 
 exports.delete = async (req, res, next) => {
     try {
-        await Comment.destroy({ where: { post_id: req.params.id } });
-        await Post.destroy({ where: { id: req.params.id } });
-        return res.status(200).json({ message: 'Publication supprimée !' })
+        const userId = verify.verifyUser(req, res, next);
+        const result = await Post.findOne({ where: { id: req.params.id }, raw: true });
+        if (userId == result.user_id) {
+            await Like.destroy({ where: { post_id: req.params.id } });
+            await Comment.destroy({ where: { post_id: req.params.id } });
+            await Post.destroy({ where: { id: req.params.id } });
+            return res.status(200).json({ message: 'Publication supprimée !' });
+        }
+        else {
+            res.status(401).json({ message: 'Utilisateur non autorisé !' });
+        }
     }
     catch (err) {
         res.status(400).json({ err });
